refactor: retire mvp/site/ and migrate surviving modules to mvp/compilers/
Move base.py (Compiler ABC), site_map.py (SiteMap), and citation_index.py
(CitationIndexGenerator) from mvp/site/ into mvp/compilers/, which is now
the canonical home for all compilation infrastructure.

Delete the XSLT/Saxon pipeline and its tests entirely:
- mvp/site/ package (strategy, pipeline, page_compiler, catalog_compiler,
  protopage_compiler, models, templates)
- src/tools/run_build.py, generate_protopages.py, render_protopages.py,
  generate_html_from_tei.py
- tests/test_strategy.py, test_compilers.py, test_pipeline.py,
  test_protopage_compilers.py

Update imports in mvp/compilers/protopage_compiler.py, run_index.py,
test_citation_index.py, and test_site_map.py. Suite: 389 passed, 0 failed.
