+++ src/python/mvp/compilers/__init__.py
@@ -1,4 +1,7 @@
+from mvp.compilers.base import Compiler, CompilationError
+from mvp.compilers.citation_index import CitationIndexGenerator
 from mvp.compilers.protopage_compiler import ProtopageChunk, ProtopageCompiler
+from mvp.compilers.site_map import SiteMap
 from mvp.compilers.transformers import (
     Family1ProseTransformer,
     SchemaRegistry,
@@ -7,10 +10,14 @@ from mvp.compilers.transformers import (
 )
 
 __all__ = [
+    "CitationIndexGenerator",
+    "Compiler",
+    "CompilationError",
     "Family1ProseTransformer",
     "ProtopageChunk",
     "ProtopageCompiler",
     "SchemaRegistry",
+    "SiteMap",
     "Transformer",
     "TransformerFactory",
 ]
