--- src/python/mvp/site/compilers/__init__.py
@@ -1,15 +1,0 @@
-from mvp.site.compilers.base import Compiler, CompilationError
-from mvp.site.compilers.catalog_compiler import CatalogCompiler, copy_static_assets
-from mvp.site.compilers.page_compiler import PageCompiler, XSLTCompiler
-from mvp.site.compilers.protopage_compiler import ProtopageCompiler, ProtopageRenderer
-
-__all__ = [
-    "Compiler",
-    "CompilationError",
-    "XSLTCompiler",
-    "PageCompiler",
-    "CatalogCompiler",
-    "copy_static_assets",
-    "ProtopageCompiler",
-    "ProtopageRenderer",
-]
