--- src/python/mvp/site/templates/chunk.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html lang="{{ chunk.language }}">
-<head>
-  <meta charset="utf-8"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1"/>
-  <title>{{ chunk.title }}</title>
-  <meta name="dc.identifier" content="{{ chunk.cts_urn }}"/>
-  <link rel="stylesheet" href="/css/reader.css"/>
-</head>
-<body class="perseus-shell">
-
-  <header class="site-header">
-    <span class="header-logo">Perseus</span>
-    <nav class="header-nav">
-      {% if chunk.prev_file %}<a href="{{ chunk.prev_file }}">&#8592; prev</a>{% endif %}
-      {% if chunk.next_file %}<a href="{{ chunk.next_file }}">next &#8594;</a>{% endif %}
-    </nav>
-  </header>
-
-  <div class="main-area">
-    <aside class="sidebar"></aside>
-    <div class="center-col">
-      <div class="passage-header">
-        <span class="passage-breadcrumb">
-          <strong>{{ chunk.title }}</strong>
-          <span style="font-size:.85em; color:#aaa; margin-left:.5em;">{{ chunk.cts_urn }}</span>
-        </span>
-        <span class="passage-nav">
-          {% if chunk.prev_file %}<a class="nav-btn" href="{{ chunk.prev_file }}">&#8592;</a>{% endif %}
-          {% if chunk.next_file %}<a class="nav-btn" href="{{ chunk.next_file }}">&#8594;</a>{% endif %}
-        </span>
-      </div>
-      <div class="text-body">
-        {% for section in chunk.sections %}
-        <div id="s{{ section.n }}" data-cts-urn="{{ section.cts_urn }}">
-          {% for para in section.paragraphs %}
-          {{ para }}
-          {% endfor %}
-        </div>
-        {% endfor %}
-      </div>
-    </div>
-    <aside class="sidebar right"></aside>
-  </div>
-
-  <footer class="site-footer">
-    <span>Perseus Digital Library</span>
-  </footer>
-
-</body>
-</html>
